+++ src/Component/Locale/BackendLanguage.php
@@ -264,7 +264,7 @@ final class BackendLanguage
         try {
             // Needed to make it possible to use the backend language in the console.
             if (defined('APPLICATION') && APPLICATION !== 'Console') {
-                Model::getContainer()->get('fork.cookie')->set('interface_language', $language);
+                Model::getContainer()->get('forkcms.cookie')->set('interface_language', $language);
             }
         } catch (RuntimeException|ServiceNotFoundException $e) {
             // settings cookies isn't allowed, because this isn't a real problem we ignore the exception