+++ src/Frontend/Core/Engine/Base/Widget.php
@@ -106,7 +106,7 @@ class Widget extends KernelLoader
         bool $addTimestamp = false
     ): void {
         // external urls always overwrite the path
-        $overwritePath = $overwritePath || $this->get('fork.validator.url')->isExternalUrl($file);
+        $overwritePath = $overwritePath || $this->get('forkcms.validator.url')->isExternalUrl($file);
 
         if (!$overwritePath) {
             $file = '/src/Frontend/Modules/' . $this->getModule() . '/Layout/Css/' . $file;
@@ -130,7 +130,7 @@ class Widget extends KernelLoader
         bool $addTimestamp = false
     ): void {
         // external urls always overwrite the path
-        $overwritePath = $overwritePath || $this->get('fork.validator.url')->isExternalUrl($file);
+        $overwritePath = $overwritePath || $this->get('forkcms.validator.url')->isExternalUrl($file);
 
         if (!$overwritePath) {
             $file = '/src/Frontend/Modules/' . $this->getModule() . '/Js/' . $file;