+++ src/Frontend/Modules/Blog/Actions/Detail.php
@@ -248,7 +248,7 @@ class Detail extends FrontendBaseBlock
         $this->form = new FrontendForm('commentsForm');
         $this->form->setAction($this->form->getAction() . '#' . FrontendLanguage::act('Comment'));
 
-        $cookie = FrontendModel::getContainer()->get('fork.cookie');
+        $cookie = FrontendModel::getContainer()->get('forkcms.cookie');
         $author = $cookie->get('comment_author');
         $email = ($cookie->has('comment_email') && filter_var($cookie->get('comment_email'), FILTER_VALIDATE_EMAIL))
             ? $cookie->get('comment_email') : null;
@@ -364,7 +364,7 @@ class Detail extends FrontendBaseBlock
     private function storeAuthorDataInCookies(array $comment): void
     {
         try {
-            $cookie = FrontendModel::getContainer()->get('fork.cookie');
+            $cookie = FrontendModel::getContainer()->get('forkcms.cookie');
             $cookie->set('comment_author', $comment['author']);
             $cookie->set('comment_email', $comment['email']);
             $cookie->set('comment_website', $comment['website']);